remove lifted file structure for HomePage
